Fixed buildtime and runtime errors under Windows 7.

+++ src/exeng/scenegraph/Mesh.cpp
@@ -319,7 +319,7 @@ namespace exeng {
         
         int Mesh::getPartCount() const {
             assert(this->impl != nullptr);
-            return this->impl->parts.size();
+            return static_cast<int>(this->impl->parts.size());
         }
